requirements.txt
#########
# grafana
#########
setuptools
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
#certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
#click==8.1.4
colorama==0.4.6
exceptiongroup==1.1.2
#fastapi==0.100.0
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
#pydantic==2.0.2
#pydantic-settings==2.0.1
#pydantic_core==2.1.2
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.7.1
#urllib3==2.0.3
#uvicorn==0.22.0
wsproto==1.2.0
#########
# report
#########
atlassian-python-api==3.41.19
certifi==2023.7.22
click==8.1.6
# cx-Oracle==8.3.0
oracledb==3.1.0
Deprecated==1.2.14
fastapi==0.101.0
influxdb==5.3.2
# lxml==4.9.3
msgpack==1.0.8
# numpy==1.25.2
oauthlib==3.2.2
pandas==2.2.3
psycopg2_binary==2.9.10
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
python-dateutil==2.8.2
requests-oauthlib==1.3.1
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
wrapt==1.15.0
pyquery==2.0.1
PyYAML==6.0.1
blinker==1.7.0
selenium-wire==5.1.0
python-dotenv==1.0.0
demjson3==3.0.6
influxdb_client==1.48.0
lxml==5.3.1
markdown==3.7
markdownify==1.1.0
jinja2
PyJWT
httpx